Add --sizeopt flag to `@nanobind` target
Controls the passing of size optimization flags for each target compiler.
Unfortunately, since this needs a `select()` based on compiler and
sizeopt yes/no, we now have to explicitly match the C++ compiler from the
configured toolchain. Also, this results in quite a few config settings
in the `@nanobind` target, which could be adjusted in visibility when
this is more polished.

The current attempt seems to build the nanobind example properly.