Remove default LTO options, dead_strip in MacOS linker options
Follows the official documentation's guidance that LTO is not strictly
necessary due to the library component. As an alternative, users can
give the `thin_lto` feature to a nanobind_extension instead.

The second dead_strip removal is due to it appearing twice in the
generated linker command. I also saw a missing symbol error during a
stubgen attempt earlier, which has since disappeared.