Deduce bindir from stubgen wrapper CWD
This makes the bindir available from Windows.

Also, harden the stub output path generation on Windows in the case that
the binary location in runfiles is not a symlink.

Also also, add a pattern file to the binary's data deps if provided,
ensuring that the file exists in the repo as an addressable target, and
is visible to the stubgen wrapper in the runfiles dir.