Expose nanobind's stubgen as a `py_binary` (#35)
* Expose nanobind's stubgen as a py_binary

Adds `typing_extensions` as a unconditional dependency
for the stubgen target.

The plan is to wrap the py_binary into a nanobind stubgen
rule that lives in the same file as all other rules.

* Add docstring, inline arg construction for py_binary

Everything is being forwarded to the py_binary that previously existed in the
top-level BUILD file.

There are still some difficulties locating the module object, since it
exists in multiple places. Maybe it too needs to be sourced from $BINDIR.

* Use template string for all files, assign labels to variables

* Move stubgen to nanobind package, depend on it in stubgen target

Stubgen does not understand file system paths to modules, so we need to
make an indirection and convert paths to nanobind extension targets to
Pythonic module paths via a Python script.

* Add stubgen wrapper script to create stubs directly in $(BINDIR)

Extracts runfiles dir and bindir from the invoker's script path,
and proceeds to put the generated stub file directly into the $(BINDIR).

What is left outstanding is to support pattern and marker files, but
that has lower prio for now.

* Deduce bindir from stubgen wrapper CWD

This makes the bindir available from Windows.

Also, harden the stub output path generation on Windows in the case that
the binary location in runfiles is not a symlink.

Also also, add a pattern file to the binary's data deps if provided,
ensuring that the file exists in the repo as an addressable target, and
is visible to the stubgen wrapper in the runfiles dir.