Fix job skip condition by contains() check (#47)
This PR changes the CI pipeline to be based on uv. It also disables abi3audit properly for all Python versions earlier than 3.12 by checking if the version from the CI job matrix is in the abi3audit candidate list.

It seems that Windows jobs take longer now than before, though, which warrants an investigation.

Also for the moment disables the recursive stubgen in the nanobind_example again with a git patch, since that hasn't landed yet.