change feature detection to look up `cc_static_library` in native module
The version gate does not work, because it is unsupported by design in Bazel.

A workaround is to use `@bazel_features`, but that requires an additional
dependency. For now, we can just try and look up `cc_static_library` from
the native module - this is consistent with `rules_cc` usage, because right
now, their `cc_static_library` is just a shim around the native rule.