fix(SABI): Use new `headers_abi3` target on Windows if available (#76)
This is a backwards-compatible fix attempt for Windows stable ABI builds.
To recap, until recently, Windows extension builds silently linked against the wrong
lib file, causing lookup errors when running a stable ABI extension on a Python version
different from the one with which the extension was built.

The fix is backwards-compatible because it decays gracefully into the old (buggy) behavior
if we do not detect the "headers_abi3" feature in `rules_python`, which marks the inclusion
of the correct ABI3 headers target.

For Mac and Linux, this is a no-op, since they never had trouble with stable ABI extensions
to begin with.