readme: Remove roadmap, add known issues section (#77)
The stable ABI linking story on Windows has now been addressed both in upstream
`rules_python` and in `nanobind-bazel`. What's left is to inform users about the
problem and how to deal with it.

In anticipation that there may be more issues in the future, add a "Known issues"
section serving as a bulletin board on the most widely reported problems with
nanobind-bazel. Let's hope there will not be too many more.