Allow shared builds with a libnanobind
Modeled after an idea on the Bazel slack. A second copy of nanobind is defined in the
internal module extension, and packaged in a `cc_shared_library` named `libnanobind`.

Changes the `nanobind_extension` to link against `libnanobind.so` if `linkstatic = False`.
This makes it possible to link multiple extensions against a single copy of nanobind.

NB: Due to the symbol resolution mechanism, `nanobind_extension`s on macOS must be given the
linker script used internally by the nanobind builds. The script portion will be refactored
into a helper routine in a subsequent commit, to be used by users of `nanobind_bazel`.