fix(linkage): Add `nanobind_link_mode` attribute to control (lib)nanobind linkage
We should not just flip `linkstatic` back and forth to request a shared build, since
that forces all other deps to be linked as shared libs, too. So instead, we add a new
attribute called `nanobind_link_mode` to the `nanobind_extension` macro, which controls
linkage of `@nanobind` for a given extension.

For convenience, the default value `nanobind_link_mode = "auto"` links nanobind based on
that value of `linkstatic`, which means nanobind extensions are by default statically linked.
Setting it to "shared" allows an explicit shared build against `libnanobind.so`.