Add new shared nanobind aliases to `@nanobind_bazel`
As for the feature request before, this forwards the internal nanobind targets
to the nanobind_bazel main module namespace.Add new shared nanobind aliases to `@nanobind_bazel`

This allows users to take more control of their targets if `nanobind_extension` and
similar don't work for them, by just using their own `rules_cc` targets and depending on
either of the forwarded nanobind flavors.

NB: To correctly link a shared libnanobind, you must depend on `@nanobind_bazel//:nanobind_shared`
in the static `deps` set, since a) that contains the library and headers, and b) it sets the
correct export macros for libnanobind.